Move high level tests to feature test
